Clarify notify with `change` condition.
Naively assumed that `change: true` meant that notifications only appear when the status changes (e.g. a branch goes from passing to failing), but based on the notifications we've been getting, it looks like it sends notifications any time the branch changes (e.g. a new build passes or fails).